Switched MenuLocation from Enum to SelectableString
